EnOcean: Refactored enocean library packages and common information (EEPIds, Parameter). Fixed dead lock in EnOcean stream reading.
