+++ bundles/binding/org.openhab.binding.enocean.test/src/test/java/org/openhab/binding/enocean/internal/bus/RockerSwitchInSwitchOnOffProfileTest.java
@@ -12,9 +12,9 @@ import static org.junit.Assert.assertEquals;
 
 import org.enocean.java.address.EnoceanId;
 import org.enocean.java.address.EnoceanParameterAddress;
+import org.enocean.java.common.EEPId;
+import org.enocean.java.common.Parameter;
 import org.enocean.java.common.values.ButtonState;
-import org.enocean.java.eep.Parameter;
-import org.enocean.java.eep.RockerSwitch;
 import org.junit.Before;
 import org.junit.Test;
 import org.openhab.core.library.items.SwitchItem;
@@ -29,7 +29,7 @@ public class RockerSwitchInSwitchOnOffProfileTest extends BasicBindingTest {
         parameterAddress = new EnoceanParameterAddress(EnoceanId.fromString(EnoceanBindingProviderMock.DEVICE_ID), CHANNEL, (String) null);
         provider.setParameterAddress(parameterAddress);
         provider.setItem(new SwitchItem("dummie"));
-        provider.setEep(RockerSwitch.EEP_ID_1);
+        provider.setEep(EEPId.EEP_F6_02_01);
         binding.addBindingProvider(provider);
     }
