+++ bundles/binding/org.openhab.binding.enocean/src/main/java/org/openhab/binding/enocean/EnoceanBindingProvider.java
@@ -9,7 +9,7 @@
 package org.openhab.binding.enocean;
 
 import org.enocean.java.address.EnoceanParameterAddress;
-import org.enocean.java.eep.EEPId;
+import org.enocean.java.common.EEPId;
 import org.openhab.binding.enocean.internal.profiles.Profile;
 import org.openhab.core.binding.BindingProvider;
 import org.openhab.core.items.Item;
