+++ bundles/binding/org.openhab.binding.enocean/src/main/java/org/openhab/binding/enocean/internal/profiles/DimmerProfile.java
@@ -8,10 +8,10 @@
  */
 package org.openhab.binding.enocean.internal.profiles;
 
+import org.enocean.java.common.Parameter;
 import org.enocean.java.common.ParameterAddress;
 import org.enocean.java.common.values.ButtonState;
 import org.enocean.java.common.values.Value;
-import org.enocean.java.eep.Parameter;
 import org.openhab.core.events.EventPublisher;
 import org.openhab.core.items.Item;
 import org.openhab.core.library.types.IncreaseDecreaseType;
